+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Advantages:
 - [endpoint]: Added createDefaultPowerSourceBatteryClusterServer(). Add Power Source cluster for a generic battery device.
 - [platform]: Added setSchema() to temporarly set the schema for the config editor.
 - [platform]: Added getSchema() to retrieve the schema from the Matterbridge plugin manager.
+- [platform]: Added return value to registerVirtualDevice().
 
 ### Changed